Refactor Model import paths across the project
- Move Model base class from `@src/core/models/base/Model` to `@src/core/domains/models/base/Model`
- Update import statements in multiple files to use the new Model path
- Ensure consistent import across test models, domain models, and other related files
