+++ src/core/domains/auth/models/ApiToken.ts
@@ -5,8 +5,8 @@ import ApiTokenObserver from '@src/core/domains/auth/observers/ApiTokenObserver'
 import UserRepository from '@src/core/domains/auth/repository/UserRepository';
 import ScopeMatcher from '@src/core/domains/auth/utils/ScopeMatcher';
 import BelongsTo from '@src/core/domains/eloquent/relational/BelongsTo';
+import Model from '@src/core/domains/models/base/Model';
 import { IModelAttributes, ModelConstructor } from '@src/core/interfaces/IModel';
-import Model from '@src/core/models/base/Model';
