+++ src/core/domains/events/models/FailedWorkerModel.ts
@@ -1,5 +1,5 @@
 import { FailedWorkerModelAttributes } from "@src/core/domains/events/interfaces/IEventWorkerConcern";
-import Model from "@src/core/models/base/Model";
+import Model from "@src/core/domains/models/base/Model";
 
 
 export interface FailedWorkerModelData extends FailedWorkerModelAttributes {}
