+++ src/tests/models/models/TestModel.ts
@@ -1,5 +1,5 @@
+import Model from "@src/core/domains/models/base/Model";
 import { IModelAttributes } from "@src/core/interfaces/IModel";
-import Model from "@src/core/models/base/Model";
 
 export interface TestModelData extends IModelAttributes {
     name: string
